Refactor Redux actions & reducers to be Ducks modules
