+++ webclient/src/components/Link/index.jsx
@@ -3,7 +3,7 @@ import { NavLink } from 'react-router-dom';
 import { connect } from "react-redux";
 import { withRouter } from 'react-router'
 import PropTypes from 'prop-types'
-import { closeSideNav } from '../../actions/sidenav';
+import { closeSideNav } from '../../redux/modules/ui/sidenav';
 
 NavLink.PropTypes = {
   to: PropTypes.string.isRequired
