+++ webclient/src/components/Toast/index.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import './style.css'
 
 import { connect } from "react-redux";
-import { hideToast } from '../../actions/toast';
+import { hideToast } from '../../redux/modules/ui/toast';
 
 class Toast extends React.Component {
   render() {
