+++ webclient/src/redux/modules/blog.js
@@ -1,6 +1,10 @@
-import { showToast } from './toast';
+import { arrayToObject, objectToIDKeyedObject, addToArrayAndSort } from '../util';
+import { showToast } from './ui/toast';
 import { push } from 'react-router-redux'
 
+/*
+ * Blog actions
+ */
 export const LIST_BLOG_POSTS = 'LIST_BLOG_POSTS';
 export const EDIT_BLOG_ENTRY = 'EDIT_BLOG_ENTRY';
 export const DELETE_BLOG_ENTRY = 'DELETE_BLOG_ENTRY';
@@ -8,6 +12,60 @@ export const LOADING_BLOG_POSTS = 'LOAD_BLOG_POSTS';
 export const BLOG_POSTS_ERRORED = 'BLOG_POSTS_ERRORED';
 export const STORE_BLOG_POST = 'STORE_BLOG_POST';
 
+/**
+ * Reducer
+ */
+let initialState = {
+  byId: {},
+  allIds: [],
+  isLoading: false,
+  hasErrored: false
+};
+
+export default function reducer(state = initialState, action) {
+  switch (action.type) {
+    case LIST_BLOG_POSTS:
+      let blogPosts = arrayToObject(action.blogPosts);
+      return {
+        ...state,
+        byId: {
+          ...blogPosts
+        },
+        allIds: Object.keys(blogPosts)
+      }
+    case DELETE_BLOG_ENTRY:
+      return {
+         ...state
+      }
+    case LOADING_BLOG_POSTS:
+      return {
+        ...state,
+        isLoading: action.isLoading
+      }
+    case BLOG_POSTS_ERRORED:
+      return {
+        ...state,
+        hasErrored: action.hasErrored,
+        isLoading: false
+      }
+    case STORE_BLOG_POST:
+      let blogPost = objectToIDKeyedObject(action.blogPost);
+      return {
+        ...state,
+        byId: {
+          ...state.byId,
+          ...blogPost
+        },
+        allIds: addToArrayAndSort(state.allIds, action.blogPost.id.toString())
+      }
+    default:
+      return state;
+  }
+}
+
+/**
+ * Action Creators
+ */
 function listBlogPosts(blogPosts){
   return {
     type: LIST_BLOG_POSTS,
