+++ webclient/src/redux/modules/ui/sidenav.js
@@ -1,0 +1,56 @@
+/*
+ * Sidenav actions
+ */
+export const OPEN_SIDENAV = 'OPEN_SIDENAV';
+export const CLOSE_SIDENAV = 'CLOSE_SIDENAV';
+export const TOGGLE_SIDENAV = 'TOGGLE_SIDENAV';
+
+ /**
+  * Reducer
+  */
+ let initialState = {
+   opened : false
+ }
+
+ export default function reducer(state = initialState, action){
+  switch (action.type) {
+    case OPEN_SIDENAV:
+      return {
+        ...state,
+        opened: true
+      }
+    case CLOSE_SIDENAV:
+      return {
+        ...state,
+        opened: false
+      }
+    case TOGGLE_SIDENAV:
+      return {
+        ...state,
+        opened: !state.opened
+      }
+    default:
+      return state;
+  }
+}
+
+/**
+ * Action Creators
+ */
+export function openSideNav(){
+  return {
+    type: OPEN_SIDENAV
+  }
+}
+
+export function closeSideNav(){
+  return {
+    type: CLOSE_SIDENAV
+  }
+}
+
+export function toggleSideNav(){
+  return {
+    type: TOGGLE_SIDENAV
+  }
+}
