+++ webclient/src/redux/rootReducer.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
 import { routerReducer as router } from 'react-router-redux'
-import blog from './blog';
-import ui from './ui';
-import auth from './auth';
+import ui from './modules/ui';
+import blog from './modules/blog';
+import auth from './modules/auth';
 
 const rootReducer = combineReducers({
   ui,
